requirements.txt
# General dependencies
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.25.0
peft>=0.5.0
tqdm>=4.66.0
numpy>=1.24.0
Pillow>=10.0.0
huggingface_hub>=0.19.0

# For dataset handling
datasets>=2.15.0
pandas>=2.0.0

# For CLIP score calculation
scikit-learn>=1.3.0

# For BLEU and ROUGE score calculation
nltk>=3.8.0
rouge-score>=0.1.2

# For stable diffusion
accelerate>=0.25.0
compel>=2.0.0

# For HTTP requests (used in PromptGenerator)
requests>=2.28.0
